Nightly build.
Moved code that handles the child proc out into its own method.

Added "backend" support proper.

If f_name is "backend", then this process is started as normal, but a
QDateTime is saved for that process. Then after the process stops, the
elapsed interval between when started and the current date-time (in UTC)
is compared (see QDateTime::secsTo()). If interval is greater or equal,
then restart.

Also, added options to grab timeout intervals from the snap config file.