Added the ecommerce plugin to the compiled set.
Updated the start script to add the lock directory if missing. Useful for programmers.
Added a few fake products for test purposes for now.
Got ecommerce.cpp to compile and the other files to pass through all the tests.
Implemented the ecommerce.js to at least add products and display them in a table.
Added buttons to delete, add, and remove items (quantity) in the cart.
Fixed the owner named "list" to "output" as it ought to be.
Fixed the feed list sort key.
Moved several taxonomy entries under content-types.
Fixed the filter parser, it was testing with TOK_IDENTIFIER instead of TOK_SEPARATOR.
Properly remove the string quotes when saving the final result.
Fixed the reading of filter arguments when the name is given by the client.
Added support for a linked_to() that supports an asterisk at the end of the path.
Properly verify that the link name, page, or type name are not empty.
Added a function to count the number of items in a map.