Added the epayment plugin.
Fixed the list of files in the ecommerce CMakeLists.txt
Renamed 'instant cart' as just plain 'Snap! Websites cart'.
Changed the eCommerceCart class so it derives from the ServerAccessTimerCallbacks.
Added tests to limit the size of the cart to (by default) 100 items x 100 quantity.
Properly implemented and made use of the eCommerceProductBase.setQuantity() function.
Started work on sending the cart content to the website so we can have a Sales Order.
Implemented a new class to time POST requests to the server so we limit the number of such requests.
Fixed the Output.DisplayOneMessage() function so it properly handles the title and body tags.
Fixed the type of the data passed as the error messages, it is a NodeList, not an Element.
Added the span tags to the title and body of messages and the missing closing message tag.