Added missing .js validations to ecommerce CMakeLists.txt
Added a POST to send the cart changes to the server.
Added a script to rebuild the cart on reload of a page.
Fixed various JS definitions that I hadn't caught before.
Added support to add new products with a quantity other than the default.
Removed the timerID variable, we never clear it so it was useless.
Changed the default time to live of the user session to 365 days instead of 5.
Added a function to get session variables without detaching them.
Fixed the detach_from_session() and made it non-const because it changes the database.