Added a version of the e-Payment PayPal plugin which is starting to get there.
Updated the e-Payment JavaScript accordingly so the PayPal payment works.
Got the cart popup to display the list of available payments, implemented a Back button and the PayPal button.
Added the "secret" table to save data that should not be accessible via the snap_expr cell() function and other such "public" features.
Updated the Editor implementation to automatically save any secret widget in the secret table instead of the revision table.
Removed the ICU dependency from the ecommerce and epayment plugins.
Properly vadlidate the locale XML/XSL files.
Fixed the locale focus and tab order definitions.
Fixed missing return in the server_access.js file.
Added an administrator view right so we can show certain pages just to administrators. We may still change these entries once we have a better view of all the Snap! rights.
Added a path to "/invoices" to save e-Commerce invoices under.
Fixed the e-Commerce handling of the cart data when reloading a page.
Got a first version generating an invoice from the cart (still missing a proper body.)
Added a security check in the cart so only allowed products can be added to the invoice.
Replaced the Update button with a Clear button (update won't be necessary.)
Fixed the tcp_client::read_line() function which would return -1 too soon if the input was missing a newline.
Make use of the safe...Value() functions in the QCassandraValue class in the snap_expr (which is accessible by end users and so often misused.)
Changed some checks from 'nullValue()' to 'size() == sizeof(...)' which is much safer.
The changes to reading value safely made the list plugin much less prone to throwing.
Added a get_current_date() in the snap_child class. In some cases, it makes more sense to use the current date rather than the start date.
Added a bio_client class which handles TCP/IP connections with TLS encryption.
Added a new class to handle plain and encrypted (TLS) HTTP requests. At this point, only the client side.
Added a dependency on OpenSSL (libssl-dev) since we need it to implement our own version of the HTTP protocol to access secure sites such as PayPal.
Added as2js as a dependency as we are now using the JSON parser.
Fixed the bug introduced by a former change to the filter and position/named parameters.
Various clean ups all over the place.