Fixed an improper recursivity in the resolution of a class, although right now we hit another case of recursivity.
This version of the code includes many std::cerr for debug purposes.
Made the is_constructor() function return the pointer to the class.
Test whether the function_node is a function before anything else.
Fixed the bug with the test of the CONSTRUCTOR attribute: we still need the function to be defined inside a class.
Fixed the function check of its children which was still using the old scheme. Now we have a clear NODE_TYPE for the type.
Added a TYPE attribute to mark nodes that are used in a type expression.
Added a cmake folder with the FindAs2Js.cmake file.
Prevent constructors from returning anything because that's forbidden in Java and C++.
Check the list of parameters for their type only if the function was not yet typed, otherwise all the parameters are already typed.
Moved the resolve_internal_type() to the compiler_package.cpp where it makes more sense.
Fixed the scripts so they are in the Native package and not the Global package.
Updated the node test to verify the new TYPE attribute.
Various clean ups.