Fixed problems with XSLT by using -1 with all toString() calls on QXmlDocument's.
Fixed the addition of attributes by adding a simplified() call on the attribute value. As far as I know two spaces within an attribute is the same as one. This only prevents JavaScript strings with spaces, but the user should use .js files.
Added the hashtag plugin. Right now it is a filter for existing tags. We want to upgrade the editor to properly handle fitlering using a background processor instead.
Fixed the avartar page XML, it was an editor-form, not a snap-form. Also added the missing session widget and removed the accesskeys which is not yet supported.
Added a filter_text() signal to the filter plugin. This can be used to further filter each text node.