Added support to specify the method of an HTTP request (i.e. PATCH instead of POST.)
Fixed a security issue that would allow a user to add any product, even those not allowed to him/her.
Fixed the OAuth2 as it required a full database (inter-process) lock.
Added the recurring payment facility for PayPal.
Added a product list to the e-Payment plugin which the e-Commerce plugin uses to return the content of its cart.
Added the code to show the list of products being purchased on PayPal.
Added the missing function set_port().
Added missing XML validations (e-Commerce, e-Payment).
Added missing dependencies (e-Commerce, e-Payment).
Added thank you, canceled, failed pages to the e-Product plugin instead of down in the PayPal implementation (since all methods of payments should use those pages.)
Added a recurring payment class that can parse a string or a compressed number in 3 parameters specifying how a payment recurs.
Moved the product type and XSLT from e-Commerce to e-Product.
Renamed products again to accomodate the page type change to use an epayment definition.
Added a couple more tests that havea recurring definition.
Redirect users from the PayPal cancel and return URLs to another page after we processed the data.
TODO: I moved the execution of a payment in the previous iterations, but we want to move it back in the on_path_execute() function. This is for standard payments.