Avoid showing a PayPal button in the list of available payments if not available (i.e. settings not available).
Added support for double (float64) in the field_search facility.
Made the e-Payment button area optional. If no buttons, do not add it at all.
Completely removed the handling of the cancel and process PayPal buttons which are not used at all (we directly go to a Thank You page).
Removed the "epayment__epayment_paypal_token: token" from the AJAX command.
Transformed the epayment::price field into a double.
Fixed the definition of the f_debug_defined flag, it was never set to true.
Documented how you make a user an administrator.