Fixed the robots.txt plugin so it reads the links properly.
Changed the product-price to a float in the e-Payment implementation as well.
Fixed the creation of the ecommerce-cart.js script to include the current object so we could move it out of the XSLT file.
Moved the header generation for the product XSTL file to e-Payment since it is part of that plugin, not e-Commerce.
Moved the CSS for the e-Payment product details from e-Commerce to e-Payment.
Fixed the product XSTL to use epayment instead of ecommerce in the path of data.
Fixed the SNAP_NAME_EPAYMENT_PRODUCT_DESCRIPTION to SNAP_NAME_EPAYMENT_PRODUCT_NAME as the field is named.