Added a test_plugin to run unit tests from the browser exactly as a client.
Added one unit test in the links plugin and another in the server_access plugin.
Fixed the is_ajax_request() function, the name of the POST variable had changed.
Fixed a crashing bug in the bio_client as the deleters were not properly specified.
Added a function to convert a date in us to a string (copy/paste in dbutils, really).
Added support for the new test_results table in dbutils.
Fixed the ecommerce "?a=view" parameter which needed to be added after the cast.
Fixed several editor functions so they now properly prevent the default click behavior.
Fixed the date used to verify the validity of a PayPal token, we cannot use the invoice date.
Various cleanups.