Found the bug in the links.cpp implementation and fixed it.
Finished up the test_plugin implementation. It works greatly.
Implemented a test to verify the multiple to multiple link create and delete functions.
Fixed and finished the unique to unique link test.
Moved the test-plugin administration page one up because these are not settings.
Added CSS to show a "waiting" background when running tests.