Implemented a first keyboard support for dropdown widgets.
Moved the hiding feature of the dropdown items from the locale timezone widget to using a class named hidden.
Allow for dropdown items to be normal, disabled, or hidden.
Added a status on dropdown items called active when using the keyboard to select an item.
Changed itemClicked() to make use of the selected_item instead of the event. That way we can easily reuse it in various places.