Fixed several function in the QHtmlSerializer to avoid converting the & of an entity that we just transformed (i.e. if we do '<' to '&lt;' first, then the '&' to '&amp;' would act on the '&' of that '&lt;'!)
Started adding support for a Snap! OAuth2 authentication.
Broke up the code checking the user login / password information from the login form so plugins other than the users plugin can log a user in the system.
Added support for the special type "plain" in editor widget to represent plain text (i.e. text saved without entities in the database).
Fixed the insert_html_string_to_xml_doc() which was not taking the & in account.
Fixed a tooltip which was copied using value-of instead of copy-of.
Fixed the taborder tag for the epayment_paypal settings.
Prevent redirects on the validate action permission check if the method is not GET or POST.
Prevent methods other than GET, HEAD, and POST to redirect the user to another page.
Prevent the use of cookies if the HTTP method is HEAD or TRACE. This also means the user cannot be logged in when using those methods.
Added an HTML error to the "method not allowed" error generated by the snap.cgi tool.
Fixed a potential bug in case the REQUEST_METHOD environment variable is not defined (it would crash).
Changed the DEBUG test to the C/C++ _DEBUG name.