Finished up the implementation of the global OAuth2 feature. The login and logout functions both work as expected.
Changed the user log in feature with an authenticated_user() function so the OAuth2 can make use of it to log the global user in.
Added a save_session() in the user_logout() in case it is called by an external function which would not otherwise save the session.
Avoid redirecting a user if it looks like it is an application using a REST API.
Now we offer a way to access the user session information. It should only be used by plugins that have an interest in handling user sessions.
Added a tool one can use to convert a username:password to base64 to test the OAuth2 and as an example of how to do it in your own apps.
Fixed the entity name being cleared by a resize(0) so we do not lose our reserved buffer.
Added a function so we can ask the snap_child class to not output cookies in the header.
Started a module named snap_software_description with the intend to implement a list of REST API, however, we need yet another plugin for that purpose.
Removed the extra < and > characters from the OAuth2 settings.