Coverity fixes.
Make sure dynamic_cast<>() returned a valid pointer.
Properly delete the network access manager object on deletion of out QHttpRequest objects.
Added a test against the network access manager pointer, if not null when about to create such an object, throw.
Check that f_reply is not null when about to be used.