Many "small" fixes (cleanup for most) from Coverity discoveries.
Corrected various variable types used with controlled_vars.
Added many controlled_vars to avoid undefined variables.
Avoid division by zero in qdomxpath, cleanly throw or return 0.0 instead.
Fixed the integer compaction of qdomxpath which used <= instead of >= in two places.
Added copy constructor and operator to the variant and atomic value of the QDomXPath to avoid copying undefined variables.
Removed an unused for(;;) in quoted_printable.
Make use of the enum controlled_vars in the snap_expr to cleanup a set of FIXME lines.
Added a couple of functions to allow char * and wchar_t * in snap_expr variables.
Fixed a few virtual destructor to make them explicitly virtual.
Fixed a few division by zero potential problems in snap_expr with an explicit throw.
Cleaned up various things in the thread (i.e. thread identifier, log why it does not start, throw if it cannot be started...)
Fixed a test in the feed plugin about a dynamic_cast<>() result that was not checked.
Added many try/catch in main() functions.
Fixed the path_error_callback class so it accepts a reference to the ipath and not a copy.
Fixed a buffer overflow bug in session_type_to_string(), the maximum had to use >=.