Mainly checking for changes in link with Coverity checks of the compiler.
Fixed some names in headers.
Bumped the copyright date to 2015, but not in all the files yet.
Switched a few parameters to use enum controlled_vars and avoid some casts.
Marked a few class destructors as virtual.
Added an os_raii.cpp/h pair to save iostream flags in an RAII manner.
Make use of this new os_raii in node_display which makes changes to the flags.
So far, found one location that needed a set_link() although that does not yet fix all the loop problems we are having with types.
Fixed a resolution parameter, resetting it to zero if set.
Verify whether a parent exists, if not throw before we start using the variable. This will work better when searching for a node of a given type.
Verify that a String() is properly defined in the contructor, otherwise throw.
Added try/catch to main() functions to avoid Coverity errors.
Assert when we call mkdir()/open() and it fails.
Removed some dead code in the lexer test. May fail now... (TBD)
Added a missing CPPUNIT_ASSERT() call.