Fixed the problem with using "Enter" under MS-Windows when logging in. It would send the POST twice.
Added method when showing the session URI on each access to the server.
Fixed the checkbox status which broke at some point in the past. It also uses code in the form.js script instead of inline.
Fixed the default button class, it would not use the correct path to the default-button/@refid.