Got multi-site users to work, although they need to "register" with their existing password, it should be different in this case though.
User registration now includes a reference from the user back to all the websites where the user registered himself.
Enhanced the register_user() function to return a status instead of true or false.
Enhanced the editor JavaScript to support sub-forms within a form.
Added editor tags to the Bare theme (needs to work differently though, we should not have to do that in the theme!)
Got the page session information in the Bare theme, that should be added by AJAX when the user choose to edit, somehow.
Removed the wrong owner for the style.css file in various themes (was forced to output.)
Fixed documentation for snapbackend deletelink action to make it work with any website in a multiwebsite installation.
Made some of the administration pages accessible through the /admin and /admin/settings pages.
Moved the user-messages styles to the output.css so it shows up as expected in all websites.