Fixed a bug in the parsing of a recurring string. I used '\0' instead of '0'.
Added a yet untested function to process a repeat payment.
Added signals to handle repeat payments.
Added a couple of fields to save Paypal answers when processing a plan payment.
Now we save the invoice number in the secret table to help in debugging.
Properly initialize a few bare pointer variables with nullptr.
Started work on snap-communicator (not compiling that code yet...)
Clearly marked the tcp_server copy constructor and operator as deleted.
Added shared pointers to udp_client and udp_server.
Renamed the recurring "fee" as a recurring "setup fee" (which it is.)
Better an error message when getting a string value from a product.