Finished the implementation of the repeat payment facility.
Changed the epayment::get_invoice_status() to return an UNKNOWN status instead of throwing. In many cases it makes it a lot easier to handle.
Added the Maximum Repeat Failures setting for repeating payment. The plan itself has also a counter but it is easier to have our own too.
Fixed the repeat payment impl function to allow more statuses to be accepted.
Fixed the lists of actions (the "list" action now prints all the actions of each website defined on your installation.)
Worked on a fix to the link functions to send signals when creating and deleting links. This way plugins that depend on certain links (i.e. permissions) can actively listen to those changes.
Updated the list so it auto-updates lists when a page's link is modified in some way.
Added an action to the list to allow someone to force lists that contain a certain page to re-run against that page.
Trying to add a try/catch to avoid losing working on certain lists when one list fails (i.e. its script is erroneous.)
Fixed the user process_login_form() function to avoid showing an error message when another plugin already added such a message.