Fixed the status of an invoice on a repeat payment when the payment does not go through.
Added an ABANDONED state for invoices. These are failed invoices that cannot be processed again.
Properly mark invoices as FAILED when some total inconsistencies are found.
Better handle the PROCESSING status in the payment of an invoice by Paypal.
Prevent repeat payments that do not have the correct state.
Fixed the default repeat failures counter to 5 instead of 0.
Fixed the loading of the repeat faiure as an int8 and not an int64.
Implemented a time limit before a following attempt in processing an auto-payment (i.e. 1 day, because with Paypal trying again instantly is not useful).
Restore invoice status to CREATED if Paypal refuses to charge the customer for a repeat payment.