Added support for the "final" keyword to define a variable.
Updated the tests so they all pass.
Updated the library as required so the new elements added while working on the compiler are all tested (100% coverage again).
Removed the wrong duplicate of the CMAKE_BUILD_TYPE from the coverage script.
Fixed the doxygen setup to get indices in alphabetical order.