Fixed the forever loop that would crash the compiler. In two places I would use an flag as is instead of the logical not of that flag.
Got the compiler class and enum tests to pass.
Fixed the resolve_name() in case of an enumeration: we used the front INUSE flag and the resolution was the child, not the id.
Fixed a problem in resolve_name(): it needed to accept NODE_MEMBER as input.
Fixed looping through the list of enumeration variables: we did not want to test the the NODE_TYPE if present.
Fixed the parsing of variables that have a type, the 'default' was not correct, we now have a NODE_TYPE.
Fixed the display of parameters. Somehow we were missing the name (string) of the parameter at some point.
Added an INUSE flag for enumarations (instead of using the variable flag).
Added a function to convert a type node to a string, however, this is totally wrong... when comparing, we should have the exact same type by pointer.
Added the + operator for strings to make it easier in various circumstances to concatenate strings.
Added tests and fixed a few internal problems.
Properly handle the return type definition when undefined: if the object is a constructor, use Void instead of Object.
Properly handle multiple 'extends' when checking for the final flag on a function member.
Finished fixing the Native packages, they now all compile in their current state.
Enhanced the compiler test so we do not have to test link destinations in full (i.e. otherwise all type definition could end up being thousands of lines.)
Various clean ups, enhancements of error messages, and optimizations.