SNAP-179: Fixed a leak bug. Make nodes attached in another node need to use weak pointers.
Transformed the "links" into specialized functions (instance, type node, attribute node, goto exit, goto enter) so we can mix shared pointers and weak pointers.
Fixed the internal_imports() test to a variable that will be set so on subsequent calls we do not re-load all the native modules.
Updated the node display function accordingly.
Removed the type_node_to_string() which is completely wrong. Instead we directly compare the pointers of the nodes representing types.
Removed the string + operator for now. We end up not using it and since we do not yet have tests...
Fixed a bug in the node test that was using LINK_max as the maximum number of variables instead of the hard coded value used in the test.