Worked on fixing the binary_operator() function which was not creating a valid function call to find the binary operator. Not completely tested yet.
Enhanced the test in the enum directive to only allow NODE_SET to have expressions.
Finally added a "clone basic node" function so we can clone nodes representing litterals.
Updated the raii_stream_flags class to also support saving the precision and width.
Force the expression() function to optimize and type the expression before it exits.
Node output now displays the number of times a node was locked.
Added a special case in the node tree handling: the replace_with() authorize the change in the parent node even if that one was locked.
Started work on an auto-numbered enumeration test. At this time it fails in the binary operator.
Fixed a potential bug in the test_as2js_node_data.ci where the expected message level and error code could be used uninitialized.
Fixed the definition of "invariant", it is not an attribute, it is a keyword defining a block in a class.
Added a file defining functions.