Added support to generate paging of a list with navigation.
Added the necessary code to calculate the number of items in a list so the navigation can be computed at once.
Fixed a bug in the read_list() function which would eventually return more than 'count' items.
Added a list action named resetlists to force Snap! to regenerate all the lists.
Added the int32 and uint32 for the dbutils. Also sorted the integers and floats in a more logical way.
Fixed the limit to the last character in char-chart to 0x110000 as it should have been.
Fixed the few display of the character chart page to use hexadecimal.
Fixed page FF in the character charts: character 0xFFFE and 0xFFFF were creating problems.
Changed the unique paypal request identifier to use the date so people can retry a payment of the same invoice multiple times.
Various updates and clean ups.