Added the necessary code to create an unsubscribe URL with a hidden email and then the tests to avoid sending emails to users. We still need to force certain emails (i.e. when registering)
Added a submit action/permission/right so that way we can POST data from a public page (i.e. for the unsubscribe, contact, survey pages)
Added a function to not redirect to the login page if someone cannot use the "submit" action.
Added better support for years before 1800 and the last_day_of_month() function. We use (year % 4 == 0) for years before 1752. We may later want to check the locale too although that function is in snap_child...
That being said, Sept 1752 throws if presented to last_day_of_month() because there are days missing within the month.
Added a function to list the resources available in a snap_child.
Added a content plugin action named dirresources to list resources of a running snap_child.
Added a couple of fields to dbutils. We still need to work on an XML file that each plugin provides instead of that ugly hack.
Fixed a bug in snap_uri that would view the sub-domains as invalid if completely empty.
Fixed the content backend reset status code so only pages of the current site get fixed and only statuses that represent an intermediate state are reset.
Fixed the problem with the lang="{$lang}" in the qrcode settings. You have to include the functions.xsl to get the declaration.
Replaced all the references to the toStdString() function with proper code.
Fixed the sendmail loop to only process the emails that match the current website key.
Added the date when creating an email because emails may or may not have the Date field. Plus that way it will be easier to deal with sorts etc.
Added a new sendmail status "unsubscribed" which is used to mark an email that a user will not be sent because he blacklisted his email.
Optimized the sendemail() function to get a row for the email status in a variable and also compute and cache the key once.
Better documented and numbered the state_t and working_t values as these are saved in the database.
Cleaned up and enhanced the editor form DTD. For example, the main auto-save and processor attributes are not used in the editor and the HTML data now includes lists. I also added float64 as a supported type (equvalent to double).
Enhanced the setPopup() function of the editor so we can ask the editor to no hide/show the Save button which is useful for forms that do no need changes (i.e. mode="save-all" and auto-save="no").
Fixed the core validation process of the editor so it actually works.
The layout plugin now checks whether a layout is valid (status of the page is NORMAL) before updating that layout.
Fixed the canonical URL, it needed to use the main_ipath instead of the plain ipath which may have changed in case of a dynamic path.
Added code to support the installation of a "-page" file from resources. This includes editor forms which the layout should not really know about. We may add a signal later.
Cleaned up the locale UnicodeString handling by creating a class which supports the QString type as input and output. Also avoid converting to and from UTF-8 since both string formats support UTF-16.
Fixed the use of various ICU objects with the use of their LocalPointer smart pointer.
Added a parse date and a parse time set of functions to the locale plugin.
Added the missing "lines" to the sizes tag, then changed that attribute to min-lines and max-lines.
Added the new min-date/max-date, min-time/max-time, and uri filters. Also implemented the validate filter.
Fixed the dropdown immediate and focus tests which still used a relative path which is wrong. This may also have been like that to work with locale multi-widget capability.
Fixed the radio button default specification on load of an unsaved widget.
Fixed the editor session to save the main path and the current path (i.e. for dynamic paths support.)
Fixed a security bug in the session plugin that would not check the used-up field if once was false.
Added helper functions to save and load parameters in the users table.
Widget validation now includes proper timezone and locale.
Fixed the exception thrown by http_cookie to make use of a snap_exception instead of std::runtime_error directly.
Various clean ups (updated/improved documentation; space before/after & and *; adding const; moving const; error messages; dead code; etc.)