Fixed the "going to box" page if we detect that the post is under "/admin/layouts/.../.../..." then we force a redirect to the referrer.
Change the name of the variable holding the form session to include a starting underscore.
Fixed a QFile load, the "qrc" (in "qrc:/...") was being removed twice.
Verify the widget name in forms to make sure they match '[_a-zA-Z][_a-zA-Z0-9]*'.
Attach the name of the widget on errors.
Cleaned up and enhanced some error messages.