SNAP-167: Added support so the logger can connect to the loggingserver if running.
Added some logging of fatal errors in the snap.cgi binary.
Fixed the extension of the log configuration files .properties.
Fixed the logger level. INHERIT does not work anymore. We now use DEBUG by default.
Added support for log_config in the snapcgi.conf file.
Better documented the logger support in the log.cpp file.
Added support for a "last logging type" so we can return to that type when reconfiguring.
Fixed the name of a few functions to better follow our syntax scheme.
Added a check to see whether the loggingserver is available or not.
Enhanced the fork_child() to setup the prctl() and start/stop the logger as required.
Removed the again counter and more flag in the snap.cgi. They were not used for a very long time.