SNAP-43: Strengthen the use of the AJAX complete callback.
Implemented the screen undarken and the display of messages directly in the AJAX library.
In terms of errors, it will display either the result.messages if it exists or result.error_message.
Made sure that all implementation of the serverAccessSuccess/Error/Complete would call the super class implementation.
Removed the many functions (serverAccessSuccess/Error/Complete) that were empty since the callbacks are all defined (none are abstract anymore.)
Use NaN to initialize a timeout identifier. The timeout setup function cannot possibly return NaN unless it really badly fail.
Fixed the ServerAccessTimer.send() function so it actually defines and make use of the lastRequest_ parameter as expected.
Updated/fixed the corresponding and other documentation.