SNAP-94: Imlemented the saving of editor fields to the draft row in case of errors.
The save function does not create a new revision if a field is invalid. Instead it saves the data in the draft row.
Fixed a bug where the QtCassandra could have cached some values that are to be saved.
Updated the documentation about the content tables: it now shows the files and secret tables.
Added support for draft and default values to be read from the "validate" JavaScript.
Added a couple of DOM helper functions.
Added a function to get the draft and suggestion keys from a path_info_t object.
Added a signal used to convert a value to a string and vice versa (need to make the other one a signal too...) That way other plugins can extend the supported types.
Make sure that the text stream used by the feed is setup for UTF-8.
Removed a few old and useless comments.
Fixed a toString() that should have used -1 for the indentation.
Added a get_user_identifier() for the current user.