SNAP-94: Implemented the load so it makes use of the last draft if available and has a *valid* date.
Added the implementation of the value_to_string function as a signal.
Renamed the value_info_t into string_to_value_info_t that way we can have a value_to_string_info_t.
Implemented the gathering of the default value from the preset tag.
Removed the "value_to_string" call for draft data because we just save those as strings anyway (because the conversion of invalid data is not unlikely to fail!)