SNAP-42: Added support for an "angry" list which really prevent all outgoing emails, contrary to the blacklist that let important emails (user account setup / password change emails) through.
Added the necessary parameter to allow user emails sent by the users plugin to be received even when the user is on our blacklist.
Strengthen the save/load parameter for a user to support any type. We now save the date when the user unsubscribes.
Added a users::modified field for the users table.
Updated the unsubscribe setup so it saves the modified and the created dates as expected.