SNAP-186: Added support for auto-reset inside editor forms. At this time this resets and disables all the widgets when the timeout occurs. It also calls the user specified callback.
Fixed the editor focus() function: we do not setup the focus on a widget which is disabled.
Fixed the enable (disable) so as to set the contenteditable to false while disabled (avoiding the focus capability.)
Added a blur() function on widgets.
Added a function to check whether a widget is enable.
Fixed various documentation entries.