Implemented the grammar. Started to write tests for it.
Added doxygen support. Documented the lexer and the basic grammar.
Fixed the syntax.txt file, although I am thinking to get rid of it because those rules appear in the Doxygen file too.
Fixed the constexpr definitions from the lexer. They need to be inline.
Added support for the lexer to directly read a '$name' as a variable.
Added support for the '&' token (REFERENCE).
Added support for the '~' token (PRECEDED--new in CSS 3).
Changed the PERCENT value to be between 0 and 1 instead of 0 and 100.
Added functions to handle children in a node.
Added a function to display a node recursively with indentation.
Added a color parser and compressed output function.