SNAP-187: Actually found out how to do the semi-colon optimization in the parser.
Added a node type named COMPONENT_VALUE to clearly distinguish those.
Keep WHITESPACE tokens, it is really required in the list of selectors.
Optimized the tree in the parser: if a list has just one element return that element instead of the list with one element.
Made the variable set in the global scope function as expected.
Updated the tests accordingly.