SNAP-187: Small cleanups.
Enhanced the console log layout.
Made the "---- new snap_child session()" an INFO log message and not just part of debug.
Added a try/catch around the do_update() calls, although somehow it does not seem like that helped in fixing the finball update problem.
Added a call to the logger each time snapinit starts. That way we know that it is happening.