SNAP-187: Started on the compiler tests verifying lists of selectors and first few validations of input data.
Added counters to the error implementation.
Added a class to protect error counters while running code.
Added a class to support changing the error output stream in an exception safe manner.
Added an exception instead of using exit(x) so that way we can write tests and a GUI can prevent exiting.
Added many operators (!=, ?:, >=, ...) and a first implementation of expression reduction.
Added a class to parse/handle An+B strings.
Added a BOOLEAN type to the node.
Added various functions to handle the children of a node.
Added support for a variable function: $varname(args).
Added a way to save variables in nodes.
Added a function to clear all the variables defined in a node.
Added the to_string() function to the node so we can output nodes in CSS.
Added the INFO and DEBUG levels for printing with @info/@message and @debug.
Applied a fix to properly allow nested @-keywords.
Applied a fix to properly handle the ';' and '!' that may mark the end of a rule.
Implemented various @-commands that are specific to the CSS preprocessor (@if, @else [if], @error, @message...).
Implemented a first version of a function to replace variables with their content in a tree of nodes.
Implemented a few functions used to find external SCSS files and load them.
Fixed the stylesheet() parser to keep a LIST at the top, without that list we cannot properly compile everything.
Fixed the '#' by itself: it is invalid and generates an error.
Changed the COMMENT handling, the lexer checks for the '@preserve' and immediately discard comments without the word.
Updated various tests accordingly. This is not finished though.
Enhanced the compiler state to include the root too (we cannot have just on root when we load sub-scripts.)
Enhanced the test with support with --scripts (paths to scripts) and --verbose (show errors in std::cerr.)
Enhanced various scripts and documents.